spec: use int64, allow counter-flow messages (#71)
* Do not use float64 when int64 would suffice

It makes sense to use float64 where the measurement unit is
a floating, but using it when we're sending integer quantities
is actually a bit surprising.

While it's true that `float64` is native to the i386 platform
and `int64` it's not, I don't even know whether `int64` is slower
than `float64` and, even if that was the case, it would probably
be weird anyway to use `float64` rather than `int64`.

Another reason why I was using `float64` is that I didn't know
how to extact `int64`s from C code. But C code is gone in the
previous commit. Conclusion: let's use `int64` instead.

Now, why `int64` and not `uint64`? The different is subtle. We
know that in many cases golang is using signed integers also
for sizes, so that is not so surprising in golang. In Java there
is no unsigned, so in theory I could say I am doing this for
the sake of good old Java. The reality is that JSON cannot really
serialize numbers larger than 2^53 as integers, therefore the
distinction betwwn `int64` and `uint64` here is academic. Thus,
I have picked the most natural feat in golang.

(Should I check for overflow? In theory. In practice I think we
are good for quite some years without checking for overflow of
`int64` variables. Also, in the server side there are checks for
overflow of `int64`, but not of `int53`. In the client side I
think we should not care, since that is a minimal client.)

* spec: correctly mention the TLS overhead

Also, we don't have UDP overhead that we know of.

* spec: be pedantic and clarify the correct JSON integer bounds

As mentioned in the spec, this is quite an academic concern for
the moment and most likely in the future. However, clarify we have
thought about this issue, we currently consider it relevant for
pedantic implementations only, and recommend pedantic code to
make sure the int53 boundary is correctly respected.

We may change our mind in the future. (The most likely course of
action is that we'll specify what a pedantic implementation is
supposed to do in such case. For now it doesn't matter.)

* spec: explain why we need counter-flow textual messages

Do not require these messages to be sent. Yet, require a party to be
prepared to receive these messages.

The reason why I'm doing this now is to take advantage of the
recent bump in the specification version.

+++ ndt7/client.go
@@ -41,7 +41,7 @@ func (cl Client) Download() error {
 	conn.SetReadLimit(spec.MinMaxMessageSize)
 	defer conn.Close()
 	t0 := time.Now()
-	num := float64(0.0)
+	num := int64(0.0)
 	ticker := time.NewTicker(minMeasurementInterval)
 	log.Info("Starting download")
 	for {
@@ -69,7 +69,7 @@ func (cl Client) Download() error {
 			}
 			break
 		}
-		num += float64(len(mdata))
+		num += int64(len(mdata))
 		if mtype == websocket.TextMessage {
 			// Unmarshal to verify that this message is correct JSON but do not
 			// otherwise process the message's content.
