+++ ndt7/model/bbrinfo.go
@@ -4,7 +4,7 @@ package model
 // The BBRInfo struct contains information measured using BBR.
 type BBRInfo struct {
 	// MaxBandwidth is the max bandwidth measured by BBR in bits per second.
-	MaxBandwidth float64 `json:"max_bandwidth"`
+	MaxBandwidth int64 `json:"max_bandwidth"`
 
 	// MinRTT is the min RTT measured by BBR in milliseconds.
 	MinRTT float64 `json:"min_rtt"`
